Begin initial reference window refactor + meta lookup
