+++ src/lib/js/tinymce/components/reference-window/components/ManualEntryContainer.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { referenceWindowEvents as LocalEvents } from '../../../utils/Constants';
+import { referenceWindowEvents as LocalEvents } from '../../../../utils/Constants';
 
 import { People } from './People';
 import { MetaFields } from './MetaFields';
