+++ src/lib/js/tinymce/components/reference-window/components/__tests__/ManualEntryContainer-test.tsx
