+++ tsconfig.json
@@ -72,18 +72,18 @@
         "src/lib/js/tinymce/components/pubmed-window/Paginate.tsx",
         "src/lib/js/tinymce/components/pubmed-window/PubmedWindow.tsx",
         "src/lib/js/tinymce/components/pubmed-window/ResultList.tsx",
-        "src/lib/js/tinymce/components/reference-window/__tests__/ButtonRow-test.tsx",
-        "src/lib/js/tinymce/components/reference-window/__tests__/IdentifierInput-test.tsx",
-        "src/lib/js/tinymce/components/reference-window/__tests__/ManualEntryContainer-test.tsx",
-        "src/lib/js/tinymce/components/reference-window/__tests__/MetaFields-test.tsx",
-        "src/lib/js/tinymce/components/reference-window/__tests__/People-test.tsx",
-        "src/lib/js/tinymce/components/reference-window/ButtonRow.tsx",
-        "src/lib/js/tinymce/components/reference-window/IdentifierInput.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/__tests__/ButtonRow-test.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/__tests__/IdentifierInput-test.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/__tests__/ManualEntryContainer-test.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/__tests__/MetaFields-test.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/__tests__/People-test.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/ButtonRow.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/IdentifierInput.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/ManualEntryContainer.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/MetaFields.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/People.tsx",
+        "src/lib/js/tinymce/components/reference-window/components/ReferenceWindow.tsx",
         "src/lib/js/tinymce/components/reference-window/index.tsx",
-        "src/lib/js/tinymce/components/reference-window/ManualEntryContainer.tsx",
-        "src/lib/js/tinymce/components/reference-window/MetaFields.tsx",
-        "src/lib/js/tinymce/components/reference-window/People.tsx",
-        "src/lib/js/tinymce/components/reference-window/ReferenceWindow.tsx",
         "src/lib/js/tinymce/index.ts",
         "src/lib/js/utils/__tests__/HelperFunctions-test.ts",
         "src/lib/js/utils/__tests__/Modal-test.ts",
